Fold method routing into ControlCommandExecutionPolicy enum
The main-thread-callable set is a subset of the worker set, so it
becomes an associated value: socketWorker(mainThreadCallable:). One
classification entry point, init(forMethod:), replaces the
ControlMethodRoutingPolicy struct; the subset relationship is now
structural instead of a second table query.
