+++ Packages/CmuxControlSocket/Sources/CmuxControlSocket/Wire/ControlCommandExecutionPolicy.swift
@@ -1,11 +1,83 @@
-/// Where a v2 control command executes (was `SocketCommandExecutionPolicy`).
+/// Where a v2 control command executes (was `SocketCommandExecutionPolicy` +
+/// the `socketWorkerV2Methods`/`mainThreadCallableSocketWorkerV2Methods`
+/// tables on `TerminalController`).
 ///
 /// An isolation-intent value: the dispatcher consults it to decide whether a
-/// method runs on the main actor or stays on the socket-worker thread.
+/// method runs on the main actor or stays on the socket-worker thread. The
+/// main-thread-callable refinement only exists for worker-lane methods, so it
+/// is an associated value of that case rather than a separate table.
 public enum ControlCommandExecutionPolicy: Sendable, Equatable {
     /// The command must run on the main actor (UI/window/workspace state).
     case mainActor
     /// The command runs on the socket-worker thread (blocking or long-running
-    /// work that must not occupy the main actor).
-    case socketWorker
+    /// work that must not occupy the main actor). `mainThreadCallable` marks
+    /// the pure, non-blocking probes that may also be invoked synchronously
+    /// from the main thread.
+    case socketWorker(mainThreadCallable: Bool)
+
+    /// Classifies a method: every `vm.`-prefixed method and the fixed
+    /// socket-worker set run on the worker; everything else runs on the main
+    /// actor.
+    ///
+    /// - Parameter method: The trimmed method name.
+    public init(forMethod method: String) {
+        if method.hasPrefix("vm.") || Self.socketWorkerMethods.contains(method) {
+            self = .socketWorker(
+                mainThreadCallable: Self.mainThreadCallableSocketWorkerMethods.contains(method)
+            )
+        } else {
+            self = .mainActor
+        }
+    }
+
+    /// True when the command runs on the socket-worker thread.
+    public var runsOnSocketWorker: Bool {
+        if case .socketWorker = self { return true }
+        return false
+    }
+
+    /// Methods that run on the socket-worker thread instead of the main actor.
+    private static let socketWorkerMethods: Set<String> = [
+        "system.ping",
+        "system.capabilities",
+        "auth.status",
+        "auth.begin_sign_in",
+        "auth.sign_out",
+        "feedback.submit",
+        "feed.push",
+        "feed.permission.reply",
+        "feed.question.reply",
+        "feed.exit_plan.reply",
+        "browser.download.wait",
+        "browser.profiles.list",
+        "browser.profiles.create",
+        "browser.profiles.rename",
+        "browser.profiles.clear",
+        "browser.profiles.delete",
+        "browser.import.cookies",
+        "mobile.attach_ticket.create",
+        "system.top",
+        "system.memory",
+        "workspace.remote.pty_sessions",
+        "workspace.remote.pty_close",
+        "workspace.remote.pty_detach",
+        "workspace.remote.pty_bridge",
+        "workspace.remote.pty_resize",
+        "sidebar.custom.validate",
+        "sidebar.custom.reload",
+        "sidebar.custom.select",
+        // debug.sidebar.simulate_drag intentionally runs on the socket worker
+        // so its Thread.sleep between drag-state ticks doesn't block the main
+        // actor (which still owns the SidebarDragState mutations via
+        // v2MainSync). Running on .mainActor would deadlock the UI for the
+        // entire simulation, defeating the profiling workload.
+        "debug.sidebar.simulate_drag",
+    ]
+
+    /// Socket-worker methods that are also safe to invoke from the main
+    /// thread (pure, non-blocking probes).
+    private static let mainThreadCallableSocketWorkerMethods: Set<String> = [
+        "system.ping",
+        "system.capabilities",
+    ]
 }
